fix matcher for registration api, change primitive data types to wrappers for rank entities and dtos, change deprecated JsonSerializer, fix tests

+++ web-api/src/test/java/com/wowguild/arguments/BossGenerator.java
@@ -10,7 +10,7 @@ public class BossGenerator {
         boss.setName(name);
         boss.setId(id);
         boss.setDifficulty(difficulty);
-        boss.setEncounterID(setEncounterID);
+        boss.setEncounterID((long) setEncounterID);
         boss.setZone(null);
 
         return boss;
